ui: Extract LabelCommon implementations into a macro
