+++ crates/ui/src/components/label/loading_label.rs
@@ -1,5 +1,5 @@
 use crate::prelude::*;
-use gpui::{Animation, AnimationExt, FontWeight};
+use gpui::{Animation, AnimationExt};
 use std::time::Duration;
 
 #[derive(IntoElement)]
@@ -18,67 +18,7 @@ impl LoadingLabel {
     }
 }
 
-impl LabelCommon for LoadingLabel {
-    fn size(mut self, size: LabelSize) -> Self {
-        self.base = self.base.size(size);
-        self
-    }
-
-    fn weight(mut self, weight: FontWeight) -> Self {
-        self.base = self.base.weight(weight);
-        self
-    }
-
-    fn line_height_style(mut self, line_height_style: LineHeightStyle) -> Self {
-        self.base = self.base.line_height_style(line_height_style);
-        self
-    }
-
-    fn color(mut self, color: Color) -> Self {
-        self.base = self.base.color(color);
-        self
-    }
-
-    fn strikethrough(mut self) -> Self {
-        self.base = self.base.strikethrough();
-        self
-    }
-
-    fn italic(mut self) -> Self {
-        self.base = self.base.italic();
-        self
-    }
-
-    fn alpha(mut self, alpha: f32) -> Self {
-        self.base = self.base.alpha(alpha);
-        self
-    }
-
-    fn underline(mut self) -> Self {
-        self.base = self.base.underline();
-        self
-    }
-
-    fn truncate(mut self) -> Self {
-        self.base = self.base.truncate();
-        self
-    }
-
-    fn single_line(mut self) -> Self {
-        self.base = self.base.single_line();
-        self
-    }
-
-    fn buffer_font(mut self, cx: &App) -> Self {
-        self.base = self.base.buffer_font(cx);
-        self
-    }
-
-    fn inline_code(mut self, cx: &App) -> Self {
-        self.base = self.base.inline_code(cx);
-        self
-    }
-}
+impl_label_common!(LoadingLabel);
 
 impl RenderOnce for LoadingLabel {
     fn render(self, _window: &mut Window, _cx: &mut App) -> impl IntoElement {
